+++ Cargo.toml
@@ -31,7 +31,12 @@ bitflags = "0.1"
 libc = "0.2"
 lazy_static = "0.1"
 rustc-serialize = { optional = true, version = "0.3" }
-serde = { optional = true, version = "0.6" }
+serde = { optional = true, version = "0.9" }
+serde_derive = { optional = true, version = "0.9" }
+
+[features]
+
+serialization = ["serde", "serde_derive"]
 
 [dependencies.tcod-sys]
 path = "tcod_sys"
@@ -39,4 +44,4 @@ version = "4.0.1"
 
 [dev-dependencies]
 rand = "0.3"
-serde_json = "0.6"
+serde_json = "0.9"
