+++ src/lib.rs
@@ -43,7 +43,10 @@
 #[macro_use] extern crate bitflags;
 #[macro_use] extern crate lazy_static;
 #[cfg(feature = "rustc-serialize")] extern crate rustc_serialize;
-#[cfg(feature = "serde")] extern crate serde;
+#[cfg(feature = "serialization")] extern crate serde;
+#[macro_use]
+#[cfg(feature = "serialization")] extern crate serde_derive;
+
 #[cfg(test)] extern crate serde_json;
 
 pub use bindings::{AsNative, FromNative};
@@ -67,7 +70,7 @@ pub mod system;
 
 mod bindings;
 mod console_macros;
-mod serde_impls;
+#[cfg(feature = "rustc-serialize")]
 mod rustc_serialize_impls;
 
 pub type RootConsole = console::Root;
